Fixes bug-601
